need to reduce the dimensionality

+++ DebugUtils.cpp
@@ -7,7 +7,7 @@ void debug_print_array(size_t M, size_t N, float *const *array)
 {
 #ifdef DEBUG
     std::cout << '[';
-    for (int row = 0; row < M; row++)
+    for (size_t row = 0; row < M; row++)
         debug_print_array(N, array[row]);
     std::cout << ']' <<std::endl;
 #endif
@@ -15,7 +15,7 @@ void debug_print_array(size_t M, size_t N, float *const *array)
 void debug_print_array(size_t N, float const *array) {
 #ifdef DEBUG
     std::cout << '[';
-    for (int col = 0; col < N; col++) {
+    for (size_t col = 0; col < N; col++) {
         DEBUG_ONLY(std::cout << array[col] << ((col+1) < N? "," : ""));
     }
     std::cout << ']' <<std::endl;
